renamed: customer service package name to ecommerce
