+++ services/customer-service/src/main/java/com/fvrvz/ecommerce/repositories/CustomerRepository.java
@@ -1,6 +1,6 @@
-package com.fvrvz.customer_service.repositories;
+package com.fvrvz.ecommerce.repositories;
 
-import com.fvrvz.customer_service.models.Customer;
+import com.fvrvz.ecommerce.models.Customer;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface CustomerRepository extends MongoRepository<Customer, String> {
