+++ services/customer-service/src/test/java/com/fvrvz/ecommerce/CustomerServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fvrvz.customer_service;
+package com.fvrvz.ecommerce;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class CustomerServiceApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
